Time: 4 ms (30.65%), Space: 46.1 MB (73.37%) - LeetHub

+++ 0004-median-of-two-sorted-arrays/0004-median-of-two-sorted-arrays.java
@@ -28,7 +28,7 @@ class Solution {
         {
             int mid1 = merged[total/2-1];
             int mid2 = merged[total/2];
-            return (mid1 + mid2) / 2.0;
+            return ((double)mid1 + (double)mid2) / 2.0;
         }
